Move Mandate base classes to mandate package
