+++ core/src/main/scala/org/scalawag/jibe/backend/Commander.scala
@@ -1,5 +1,6 @@
 package org.scalawag.jibe.backend
 
+import org.scalawag.jibe.mandate.MandateExecutionContext
 import org.scalawag.jibe.mandate.command.{BooleanCommand, Command, UnitCommand}
 
 import scala.util.{Failure, Success}
