+++ core/src/main/scala/org/scalawag/jibe/mandate/CompositeMandate.scala
@@ -1,4 +1,6 @@
-package org.scalawag.jibe.backend
+package org.scalawag.jibe.mandate
+
+import org.scalawag.jibe.backend.Commander
 
 abstract class CompositeMandateBase(override val description: Option[String], val mandates: Seq[Mandate]) extends Mandate
