+++ core/src/main/scala/org/scalawag/jibe/mandate/MandateExecutionContext.scala
@@ -1,7 +1,6 @@
-package org.scalawag.jibe.backend
+package org.scalawag.jibe.mandate
 
-import java.io.File
-
+import org.scalawag.jibe.backend.Commander
 import org.scalawag.timber.api.Logger
 
 // This will make it easier to add more capabilities to the execution context without having to rewrite all existing
