+++ core/src/main/scala/org/scalawag/jibe/mandate/Resource.scala
@@ -1,4 +1,4 @@
-package org.scalawag.jibe.backend
+package org.scalawag.jibe.mandate
 
 trait Resource
