+++ core/src/main/scala/org/scalawag/jibe/mandate/TestMandates.scala
@@ -1,7 +1,5 @@
 package org.scalawag.jibe.mandate
 
-import org.scalawag.jibe.backend.{MandateExecutionContext, MandateHelpers, StatelessMandate}
-
 case object NoisyMandate extends StatelessMandate {
   override val description = Some("Make a lot of noise.")
